requirements.txt
colorama >= 0.4.3
simple-term-menu >= 1.2.1
pyhmy >= 20.5.20
python-dotenv >= 0.8.0
psutil >=5.8.0